+++ lib/metro/stubs/provider_stub.dart
@@ -1,0 +1,16 @@
+import 'package:recase/recase.dart';
+
+String providerStub(ReCase rc) => '''
+import 'package:nylo_framework/nylo_framework.dart';
+
+class ${rc.pascalCase}Provider implements NyProvider {
+
+  boot(Nylo nylo) async {
+   
+     // boot your provider
+     // ...
+   
+     return nylo;
+  }
+}
+''';